ocr-project-backend/requirements.txt
fastapi
fastapi-users[sqlalchemy]
uvicorn[standard]
asyncpg
alembic
redis
celery
pyOpenSSL==22.1.0
boto3
google-cloud==0.34.0
jmespath==1.0.0
pdf2image==1.16.0
Pillow==9.1.1
python-dateutil==2.8.2
reportlab==3.6.9
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.9
psycopg2==2.9.3
lxml==4.9.0
protobuf==4.21.1
testresources
google-cloud-vision
google-cloud-storage
python-bidi
requests
asgiref==3.5.2
filetype==1.2.0
python-dotenv==0.20.0
sqlalchemy-citext==1.8.0
pdf2image==1.16.0
aiohttp==3.8.4
elasticsearch==8.6.2
arabic-reshaper==3.0.0
